+++ shared/src/commonMain/kotlin/com/dummyshop/shared/data/remote/dto/ProductDto.kt
@@ -7,7 +7,7 @@ data class ProductDto(
     val id: Long,
     val title: String,
     val description: String = "",
-    val price: Int,
+    val price: Double,
     val category: String,
     val thumbnail: String? = null
 )
